requirements.txt
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
Flask==1.1.2
future==0.18.2
get==2019.4.13
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
line-bot-sdk==1.16.0
MarkupSafe==1.1.1
post==2019.4.13
public==2019.4.13
python-firebase==1.2
query-string==2019.4.13
request==2019.4.13
requests==2.23.0
urllib3==1.25.8
Werkzeug==1.0.1
